Make Timestamp more generic
